pyproject.toml
[project]
name = "mcp-financial-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.30",
    "crewai>=1.8.0",
    "crewai-tools>=1.8.0",
    "fastmcp>=2.12.5",
    "matplotlib>=3.10.8",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
]
